Url syntax verification
The verification is made using a the laravel rule: 'active_url' and a
regex pattern. The rule checks whether the text represents a url and
whether the url is active. The regex checks whether the entered url is
the one expected. Such if we expect a youtube url, then the entered url
must be a youtube one and not just any other.

modified:   .gitignore
modified:   app/Http/Controllers/ArtistsController.php
modified:   resources/views/artists/edit-profile/your-works.blade.php

Fixes #15